requirements.txt
fastapi[standard]
pymongo
python-dotenv
pymongo[encryption]

